All styles files added to the styles folder
