+++ screens/SplashScreen/SplashScreen.js
@@ -1,9 +1,7 @@
 import React from 'react'
-import { View, Text, StyleSheet, Image, Button } from 'react-native'
+import { View, Text, Image, Button } from 'react-native'
 
-
-
-
+import styles from '../../styles/SplashScreen/SplashScreen';
 
 const SplashScreen = props => {
 
@@ -16,44 +14,5 @@ const SplashScreen = props => {
         </View>
     )
 }
-
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        backgroundColor: '#3170F5',
-
-  
-    },
-
-    logo: {
-        resizeMode: 'stretch',
-        height: 100,
-        width: 100
-    },
-    
-    platanusTitle: {
-        fontSize: 20,
-        color: 'white'
-        
-    },
-
-    lakatamTitle:{
-        fontSize: 40,
-        fontStyle: 'italic',
-        color: 'white'
-    },
-
-    button: {
-        backgroundColor: '#FAC537',
-        borderColor: 'white',
-        borderRadius: 10,
-        marginTop: 30,
-    }
-    
-})
-
 
 export default SplashScreen
