+++ styles/NewTeamScreen/NewTeamScreen.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import color from '../../styles/colors';
+import color from '../colors';
 
 const styles = StyleSheet.create({
   container: {
@@ -11,18 +11,21 @@ const styles = StyleSheet.create({
     elevation: 8,
     borderRadius: 5,
     backgroundColor: color.white,
-    paddingTop: '25%',
+    paddingTop: '10%',
   },
-  formCard: {},
+  formCard: {
+
+  },
   input: {
     width: '90%',
     marginHorizontal: '10%',
+    margin: "1%"
   },
   title: {
     textAlign: 'center',
     fontSize: 20,
     fontWeight: 'bold',
-    paddingBottom: '10%',
+    paddingBottom: '3%',
   },
   tag: {
     marginTop: '10%',
@@ -33,12 +36,22 @@ const styles = StyleSheet.create({
     height: 40,
     borderBottomWidth: 0.5,
     fontSize: 15,
+    marginBottom: "2%"
   },
   buttonContainer: {
     flexDirection: 'row',
     width: '100%',
     justifyContent: 'space-between',
     padding: '10%',
+  },
+  addUser: {
+    width: '90%',
+    height: 60,
+    borderRadius: 5,
+    margin: "5%",
+    flexDirection: 'column',
+    justifyContent: 'center',
+    fontSize: 30,
   },
   cancelButton: {
     backgroundColor: color.gray,
